Move document.py to main.py to use macros
