--- docs/tables.template.md
@@ -1,7 +1,0 @@
-# Data Table Schemas
-
-Data table schemas are specified in JSON and are compatible with the
-[frictionless data](https://specs.frictionlessdata.io/table-schema/) table
-schema standards.
-
-{TABLES}
